Refactors project structure
Restructures the project by grouping modules into logical packages
(api, cli, core, exporters). This improves code organization,
maintainability, and scalability. Also includes a requirements.txt
file that documents the project's dependencies.
