+++ src/core/processors.py
@@ -5,7 +5,7 @@ import logging
 from datetime import datetime
 from typing import Dict, List, Any, Optional
 
-from team_mapper import TeamMapper
+from src.core.team_mapper import TeamMapper
 
 
 class FixturesDataProcessor:
